requirements.txt
streamlit
pandas
autogluon.tabular==0.8.2
scikit-learn
numpy
